# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Leave Management System API
  description: "REST API for Leave Management with Approval Chains, Delegations, and\
    \ Escalations"
  contact:
    name: Hackathon Team
    email: team@co.com
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- BearerAuth: []
...
paths:
...
  /api/leaves:
    post:
      tags:
      - Leaves
      summary: Submit leave request
      description: "Applies for a new leave, reserves pending balance, and initializes\
        \ the approval chain"
      operationId: applyLeave
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveApplyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeaveResponseDto"
  /api/leaves/{id}/reject:
    post:
      tags:
      - Leaves
      summary: Reject leave request step
      description: "Rejects the leave request, records reason/comment, and releases\
        \ pending balance"
      operationId: rejectLeave
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeaveResponseDto"
  /api/leaves/{id}/cancel:
    post:
      tags:
      - Leaves
      summary: Cancel leave request
      description: Cancels an existing leave request and restores pending or used
        leave balance
      operationId: cancelLeave
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeaveResponseDto"
  /api/leaves/{id}/approve:
    post:
      tags:
      - Leaves
      summary: Approve leave request step
      description: Records approval for current step and advances the workflow or
        completes final approval
      operationId: approveLeave
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalDecisionRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeaveResponseDto"
  /api/leaves/preview:
    post:
      tags:
      - Leaves
      summary: Preview leave request
      description: "Calculates working days, checks balances, and evaluates team conflict\
        \ threshold before submission"
      operationId: previewLeave
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeavePreviewRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeavePreviewResponse"
  /api/holidays:
    get:
      tags:
      - Holidays
      summary: Get official holidays
      description: Retrieves the list of scheduled public holidays used for working
        day calculations
      operationId: getHolidays
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HolidayDto"
    post:
      tags:
      - Holidays
      summary: Add public holiday
      description: Creates a new public holiday (restricted to HR administrators)
      operationId: createHoliday
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HolidayCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/HolidayDto"
  /api/demo/simulate-timeout/{id}:
    post:
      tags:
      - Demo
      summary: Simulate escalation timeout
      description: Forces an active PENDING_MANAGER leave request past its SLA to
        trigger automatic escalation to HR
      operationId: simulateTimeout
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DemoActionResponse"
...
  /api/delegations:
    get:
      tags:
      - Delegations
      summary: Get active delegations
      description: Retrieves active delegation rules created by or assigned to the
        current manager
      operationId: getDelegations
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DelegationDto"
    post:
      tags:
      - Delegations
      summary: Create delegation
      description: Delegates manager approval authority to a peer or teammate for
        a date range
      operationId: createDelegation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DelegationCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DelegationDto"
...
  /api/leaves/{id}:
    get:
      tags:
      - Leaves
      summary: Get leave request by ID
      description: Retrieves a specific leave request's details
      operationId: getLeaveById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeaveResponseDto"
  /api/leaves/{id}/timeline:
    get:
      tags:
      - Leaves
      summary: Get leave request timeline
      description: Fetches the full timeline including approval steps and audit trail
        for a leave request
      operationId: getLeaveTimeline
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LeaveTimelineDto"
...
  /api/hr/requests:
    get:
      tags:
      - HR
      summary: Get HR leave requests
      description: "Retrieves organization-wide leave requests, escalations, and pending\
        \ HR reviews with filtering"
      operationId: getHrRequests
      parameters:
      - name: status
        in: query
        description: Filter by leave status
        required: false
        schema:
          type: string
          enum:
          - PENDING_MANAGER
          - ESCALATED
          - PENDING_HR
          - APPROVED
          - REJECTED
          - CANCELLED
      - name: type
        in: query
        description: Filter by leave type code
        required: false
        schema:
          type: string
      - name: from
        in: query
        description: Filter leaves starting from date
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        description: Filter leaves ending up to date
        required: false
        schema:
          type: string
          format: date
      - name: q
        in: query
        description: Search query by employee name or reason
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: "Sort expression e.g. createdAt,desc"
        required: false
        schema:
          type: string
          default: "createdAt,desc"
      - name: page
        in: query
        description: Page number (0-indexed)
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageResponseDtoLeaveResponseDto"
  /api/calendar/team:
    get:
      tags:
      - Calendar
      summary: Get team leave calendar
      description: Retrieves leaves scheduled across team members for a given month
        (YYYY-MM)
      operationId: getTeamCalendar
      parameters:
      - name: month
        in: query
        description: Year and month in format YYYY-MM e.g. 2026-10
        required: false
        schema:
          type: string
        example: 2026-10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TeamCalendarDto"
...
  /api/delegations/{id}:
    delete:
      tags:
      - Delegations
      summary: Revoke delegation
      description: Deactivates and revokes an existing approval authority delegation
      operationId: revokeDelegation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    LeaveApplyRequest:
      required:
      - fromDate
      - leaveTypeId
      - reason
      - toDate
      type: object
      properties:
        leaveTypeId:
          type: integer
          format: int64
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        reason:
          type: string
    LeaveResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        employeeEmail:
          type: string
        leaveTypeId:
          type: integer
          format: int64
        leaveTypeCode:
          type: string
        leaveTypeName:
          type: string
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        days:
          type: integer
          format: int32
        reason:
          type: string
        status:
          type: string
          enum:
          - PENDING_MANAGER
          - ESCALATED
          - PENDING_HR
          - APPROVED
          - REJECTED
          - CANCELLED
        flagged:
          type: boolean
        flagReason:
          type: string
        createdAt:
          type: string
          format: date-time
        lastActionAt:
          type: string
          format: date-time
    ApprovalDecisionRequest:
      type: object
      properties:
        comment:
          type: string
    LeavePreviewRequest:
      required:
      - fromDate
      - leaveTypeId
      - toDate
      type: object
      properties:
        leaveTypeId:
          type: integer
          format: int64
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
    LeavePreviewResponse:
      type: object
      properties:
        workingDays:
          type: integer
          format: int32
        currentBalance:
          type: number
          format: double
        balanceAfter:
          type: number
          format: double
        conflictFlagged:
          type: boolean
        flagReason:
          type: string
        holidayDates:
          type: array
          items:
            type: string
    HolidayCreateRequest:
      required:
      - date
      - name
      type: object
      properties:
        date:
          type: string
          format: date
        name:
          type: string
    HolidayDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date
        name:
          type: string
    DemoActionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
    DelegationCreateRequest:
      required:
      - delegateId
      - fromDate
      - toDate
      type: object
      properties:
        delegateId:
          type: integer
          format: int64
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
    DelegationDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        delegatorId:
          type: integer
          format: int64
        delegatorName:
          type: string
        delegateId:
          type: integer
          format: int64
        delegateName:
          type: string
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        active:
          type: boolean
...
    PageResponseDtoLeaveResponseDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/LeaveResponseDto"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        last:
          type: boolean
    ApprovalStepDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        requestId:
          type: integer
          format: int64
        stage:
          type: string
          enum:
          - MANAGER
          - HR
        assigneeId:
          type: integer
          format: int64
        assigneeName:
          type: string
        delegatedFromId:
          type: integer
          format: int64
        delegatedFromName:
          type: string
        dueAt:
          type: string
          format: date-time
        decision:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
          - ESCALATED
        decidedAt:
          type: string
          format: date-time
        comment:
          type: string
    AuditEventDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        requestId:
          type: integer
          format: int64
        actorId:
          type: integer
          format: int64
        actorName:
          type: string
        action:
          type: string
        fromStatus:
          type: string
          enum:
          - PENDING_MANAGER
          - ESCALATED
          - PENDING_HR
          - APPROVED
          - REJECTED
          - CANCELLED
        toStatus:
          type: string
          enum:
          - PENDING_MANAGER
          - ESCALATED
          - PENDING_HR
          - APPROVED
          - REJECTED
          - CANCELLED
        comment:
          type: string
        at:
          type: string
          format: date-time
    LeaveTimelineDto:
      type: object
      properties:
        leave:
          $ref: "#/components/schemas/LeaveResponseDto"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalStepDto"
        auditEvents:
          type: array
          items:
            $ref: "#/components/schemas/AuditEventDto"
    TeamCalendarDto:
      type: object
      properties:
        teamId:
          type: integer
          format: int64
        teamName:
          type: string
        month:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/TeamLeaveEventDto"
    TeamLeaveEventDto:
      type: object
      properties:
        leaveId:
          type: integer
          format: int64
        employeeId:
          type: integer
          format: int64
        employeeName:
          type: string
        fromDate:
          type: string
          format: date
        toDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - PENDING_MANAGER
          - ESCALATED
          - PENDING_HR
          - APPROVED
          - REJECTED
          - CANCELLED
        leaveTypeCode:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      name: BearerAuth
      scheme: bearer
      bearerFormat: JWT